// FIL.Web.Admin/ClientApp/models/EventCreation/EventCreationViewModel.ts
export class EventCreationViewModel {
    id:number;
    altId: string;
    title: string;
    location: string;
    placename: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    country : string;
    zip: string;
    street : string;
    description: string;
    history:string;
    highlights: string;
    immersiveExperience: string;
    archdetail: string;
    archdetailId: number;
    historyId: number;
    immersiveExperienceId: number;
    highlightsId: number;
    tilesSliderImages: any;
    descpagebannerImage: any;
    inventorypagebannerImage:any;
    galleryImages: any;
    placemapImages: any;
    timelineImages: any;
    immersiveexpImages: any;
    archdetailImages: any;
    metatags: string;
    metatitle: string;
    metadescription: string;
    categoryid: number;
    subcategoryid: number;
    userAltId: string;
    amenityId: string;
    isEdit: boolean;
    hourTimeDuration: string;
    minuteTimeDuration: string;
    lat: string;
    long: string;
    tagIds: string;
}